Updated a few errors discovered last night Needed to change CUDALucas.cu line 2786 & 3028 so it will compile in MSVS 2010 for CUDA 4.2 and CUDA 5.0.

+++ trunk/CUDALucas.cu
@@ -2783,7 +2783,7 @@ check (int q)
   int error_flag;
   int error_reset = 1;
   int end = (q + 31) / 32;
-  float error_limit = (g_el - g_el / 8.0 * log (g_ei)/ log(100.0)) / 100.0;
+  float error_limit = (g_el - g_el / 8.0 * log ( (float) g_ei)/ log(100.0)) / 100.0;
   signal (SIGTERM, SetQuitting);
   signal (SIGINT, SetQuitting);
   do
@@ -3025,7 +3025,7 @@ check (int q)
       }
 	    else if (interact_result & 4)
       {
-        error_limit = (g_el - g_el / 8.0 * log (g_ei)/ log(100.0)) / 100.0;
+        error_limit = (g_el - g_el / 8.0 * log ( (float) g_ei)/ log(100.0)) / 100.0;
         //printf("%0.5f\n", error_limit);
       }
       interact_result = 0;
